requirements.txt
tensorflow-cpu==2.12.0
imutils==0.5.4
imageio==2.31.0
opencv-python==4.7.0.72
streamlit>=1.28.0
ultralytics>=8.0.200
face_recognition>=1.3.0
dlib>=19.24.1
numpy==1.24.3
folium>=0.14.0
streamlit-folium>=0.15.0
pandas>=2.0.0
twilio>=8.0.0
geopy>=2.4.0
requests>=2.31.0